Update focus state z-index for checkboxes
@thomasjbradley

I'm not looking to troubleshoot this specifically for my pattern library (primarily because it's held together by tears, duct tape, and vodka-- a true coding disaster of a project), but I often encounter issues when tabbing though my sites where the focus boxes end up behind other content. Is there a good way to avoid that, generally? Or is updating z-index the best bet?